+++ prestashop/ntresellerclub/classes/admin/NtRcAdminDataProviderInterface.php
@@ -1,0 +1,9 @@
+<?php
+if (!defined('_PS_VERSION_')) {
+    exit;
+}
+
+interface NtRcAdminDataProviderInterface
+{
+    public function getSummary();
+}